+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # CHANGELOG
 
+### 1.1.1 (2018-10-18)
+Features:
+* **Predictive Pricing:** Introduced services to handle the evaluation of prediction models
+* **LineItem:** Line items can now determine their value by evaluating a prediction model if configured to do so. A **predicted value** will be used if the `predictive_pricing` bid variable is `true` and the line item's value cannot be computed by evaluating it's dependencies. The computed value may be overriden by setting the `use_computed` bid variable to `false` or `lineItem.useComputedValueWhenAvailable` to `false`. The bid must have a non-zero value for Watts.
+* **LineItem:** `LineItem.isWeighted` determine or set a line items weighted switch. Predicted line items can optionally be weighted by their total historic inclusion count
+* **LineItem:** `LineItem.getPredicted('cost')` determines the predicted cost of the line item using its prediction models
+* **Bid|LineItem|Component:** `Bid|LineItem|Component.isPredicted(property)` determines if a property is being predicted
+* **BidEntity:** `BidEntity.hasNullDependency(property)` determines if the property depends on an undefined dependency
+
+Improvements:
+* **BidVariable:** Expose `isReserved` property
+* **PredictionModels Repository:** `repositories.predictionModels` retrieves prediction models for account or specified lineItemDefId
+* **Project:** `predictionModels` property allows access to project's prediction models
+
+Bug Fixes:
+* **BidVariable:** execute a full reassessment when `markup_strategy` is changed
+* **Bid:** `bid.entities.getDependencyValue` now returns `null` instead of `0` for undefined dependencies
+
 ### 1.0.39 (2018-08-30)
 Bug Fixes:
 * **Project:** Improve tracking of changes to the data while project is saving